Authorization-added

+++ main.py
@@ -6,13 +6,13 @@ from typing import Annotated, Optional
 from models import Todos
 from pydantic import BaseModel, Field
 from fastapi.responses import JSONResponse
-from router import auth
+from router import auth,admin
 from router.auth import get_current_user
 
 app = FastAPI()
 models.Base.metadata.create_all(bind=engine)
 app.include_router(auth.router)
-
+app.include_router(admin.router)
 
 class Todo(BaseModel):
     id: int
@@ -42,14 +42,22 @@ user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 @app.get("/")
-def get_todos(db: db_dependency):
-    return db.query(Todos).all()
+def get_todos(db: db_dependency, user: user_dependency):
+    if user is None:
+        raise HTTPException(status_code=404, detail="User not found!")
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 
 @app.get("/todo/{todo_id}")
-def get_todos(db: db_dependency, todo_id: int):
-
-    specific_todo = db.query(Todos).filter(Todos.id == todo_id).first()
+def get_todos_by_id(user: user_dependency, db: db_dependency, todo_id: int):
+    if user is None:
+        raise HTTPException(status_code=404, detail="Failed Authentication!")
+    specific_todo = (
+        db.query(Todos)
+        .filter(Todos.owner_id == user.get("id"))
+        .filter(Todos.id == todo_id)
+        .first()
+    )
     if specific_todo is None:
         raise HTTPException(status_code=404, detail="Todo not found!!")
     else:
@@ -59,16 +67,24 @@ def get_todos(db: db_dependency, todo_id: int):
 @app.post("/create")
 def create_todos(user: user_dependency, db: db_dependency, newTodo: Todo):
     if user is None:
-        raise HTTPException(status_code=404, detail="User not found!")
+        raise HTTPException(status_code=404, detail="Failed Authentication!")
     todo_model = Todos(**newTodo.model_dump(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
-
+    return {"message": "Todo Created successfully"}
 
 @app.put("/update/{todo_id}")
-def update_todos(db: db_dependency, todo_id: int, update_todo: UpdateTodo):
-
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+def update_todos(
+    user: user_dependency, db: db_dependency, todo_id: int, update_todo: UpdateTodo
+):
+    if user is None:
+        raise HTTPException(status_code=404, detail="Failed Authentication!")
+    todo = (
+        db.query(Todos)
+        .filter(Todos.owner_id == user.get("id"))
+        .filter(Todos.id == todo_id)
+        .first()
+    )
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found!!")
     update_data = update_todo.model_dump(exclude_unset=True)
@@ -80,11 +96,12 @@ def update_todos(db: db_dependency, todo_id: int, update_todo: UpdateTodo):
 
 
 @app.delete("/delete/{todo_id}")
-def delete_todos(db: db_dependency, todo_id: int):
-
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+def delete_todos(user: user_dependency,db: db_dependency, todo_id: int):
+    if user is None:
+        raise HTTPException(status_code=404, detail="Failed Authentication!")
+    todo = db.query(Todos) .filter(Todos.owner_id == user.get("id")).filter(Todos.id == todo_id).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found!!")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos) .filter(Todos.owner_id == user.get("id")).filter(Todos.id == todo_id).delete()
     db.commit()
     return {"message": "Todo Deleted successfully"}
